requirements.txt
PySide6==6.6.0
Pillow==10.1.0
bcrypt==4.1.1
pyqrcode==1.2.1
pypng==0.20220715.0
python-docx==1.1.0
pywin32
rapidfuzz>=3.0.0
numpy<2
rembg>=2.0.50
onnxruntime==1.18.1
